Comply with dependency injection principle
-Abstract factory is used to handle the creation of "volatile concrete object"

-The diagram shows the structure : 
The Application uses the `ConcreteImpl` through the Service interface. However, the Application must somehow create instances of the `ConcreteImpl`. To achieve this without creating a source code dependency on the `ConcreteImpl`, the Application calls the `makeSvc` method of the `ServiceFactory` interface. This method is implemented by the `ServiceFactoryImpl` class, which derives from `ServiceFactory`. That implementation instantiates the `ConcreteImpl` and returns it as a Service.